+++ packages/abstract-signer/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "abstract-signer/5.0.0-beta.129";
+export declare const version = "abstract-signer/5.0.0-beta.130";